backend/requirements.txt
# TIME BLOCK Backend Requirements
# Django + DRF + PostgreSQL Stack

# Core Framework
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# CORS
django-cors-headers==4.3.1

# Environment Variables
python-decouple==3.8

# OAuth 2.0
google-auth==2.27.0
requests==2.31.0

# Image Processing (for profile images, heatmap)
Pillow==12.0.0

# Data Visualization (for GitHub-style heatmap)
matplotlib==3.10.8
numpy==2.4.0
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
kiwisolver==1.4.9
pyparsing==3.2.5

# Testing
pytest==7.4.4
pytest-django==4.7.0

# Code Quality
black==24.1.1
flake8==7.0.0

# Production Server
gunicorn==21.2.0

# Utilities
python-dateutil==2.8.2
